refactor(ui): split SettingsModal.tsx into 13 focused modules (1580→178 lines)
Extract tab components, sub-components, shared constants, and a custom
hook into src/renderer/components/settings-modal/ subfolder.

Modules extracted:
- settings-modal-shared.ts: constants, utilities, SettingsModalProps
- useSettingsSync.ts: all sync/auth/password state and handlers
- SyncStatusSection.tsx: sync status display
- DisconnectConfirmButton.tsx: confirm-to-disconnect pattern
- HubDisplayNameField.tsx: Hub display name input
- SyncDataResetSection.tsx: remote data scan and deletion
- LocalDataResetGroup.tsx: local data reset UI
- PasswordSection.tsx: encryption password form
- SettingsToolsTab.tsx: theme, language, defaults, security
- SettingsDataTab.tsx: Google auth, sync controls, Hub settings
- SettingsNotificationTab.tsx: notification list
- SettingsTroubleshootingTab.tsx: sync/local data reset, import/export
- SettingsModal.tsx: orchestrator (modal shell + tab bar)

Original path re-exports via barrel for backward compatibility.
